xtract CanvasPeopleAccessPanel from CanvasPage
